pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "evo"
authors = [{name = "Philippa Liggins", email = "pkl28@cam.ac.uk"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3"]
dynamic = ["version", "description"]
requires-python = ">=3.7"

dependencies = ["numpy>=1.21.0",
  "scipy>=1.7.0",
  "matplotlib>=3.4.2",
  "mpmath>=1.2.1",
  "pyyaml",
  "ruamel.yaml",]

[project.urls]
Home = "https://github.com/pipliggins/EVo"
